+++ TankProject/TankProject/MultiplayerGameState.hpp
@@ -29,7 +29,7 @@ public:
 
 private:
 	void						updateBroadcastMessage(sf::Time elapsedTime);
-	void						handlePacket(sf::Int32 packetType, sf::Packet& packet);
+	void						handlePacket(sf::Int8 packetType, sf::Packet& packet);
 
 
 private:
@@ -43,7 +43,7 @@ private:
 
 	Tank						*playerTank, *secondPlayerTank;
 	std::map<int, PlayerPtr>	mPlayers;
-	std::vector<sf::Int32>		mLocalPlayerIdentifiers;
+	std::vector<sf::Int8>		mLocalPlayerIdentifiers;
 	sf::TcpSocket				mSocket;
 	bool						mConnected;
 	std::unique_ptr<GameServer> mGameServer;
